Move flag to helpers
