+++ vessel/utils/unified_diff.py
@@ -33,7 +33,7 @@ from typing import Any, Optional
 import portion  # type: ignore
 
 from vessel.diff.helpers.diffline import DiffLine
-from vessel.utils.flag import Flag
+from vessel.diff.helpers.flag import Flag
 
 logger = getLogger(__name__)
